Migrated two problems in Linked List before changing the strategy
