+++ coollime/src/main/java/coollime/linkedlist/checkiflinkedlisthasacycyle/Solution.java
@@ -1,0 +1,22 @@
+package coollime.linkedlist.checkiflinkedlisthasacycyle;
+
+import coollime.helper.ListNode;
+
+public class Solution {
+
+  public boolean hasCycle(ListNode head) {
+    if (head == null || head.next == null) {
+      return false;
+    }
+    ListNode slow = head;
+    ListNode fast = head.next;
+    while (fast != null && fast.next != null) {
+      if (slow == fast) {
+        return true;
+      }
+      slow = slow.next;
+      fast = fast.next.next;
+    }
+    return false;
+  }
+}
