+++ docs/linkedlist/InsertInSortedLinkedList.md
@@ -1,72 +1,37 @@
-<!----- Conversion time: 1.733 seconds.
-
-
-Using this Markdown file:
-
-1. Cut and paste this output into your source file.
-2. See the notes and action items below regarding this conversion run.
-3. Check the rendered output (headings, lists, code blocks, tables) for proper
-   formatting and use a linkchecker before you publish this page.
-
-Conversion notes:
-
-* GD2md-html version 1.0β13
-* Tue Jan 15 2019 19:55:45 GMT-0800 (PST)
-* Source doc: https://docs.google.com/open?id=1oKTaFeh19icFXuNC2stf5PAeT7-hVgFhF9UzwIFKC1s
------>
-
-
-
 # Insert In Sorted Linked List
 
 [https://app.laicode.io/app/problem/39](https://app.laicode.io/app/problem/39)
 
-
 ## Description
-
 
     Insert a value in a sorted linked list.
 
 
     Examples
 
-
-
-*   L = null, insert 1, return 1 -> null
-*   L = 1 -> 3 -> 5 -> null, insert 2, return 1 -> 2 -> 3 -> 5 -> null
-*   L = 1 -> 3 -> 5 -> null, insert 3, return 1 -> 3 -> 3 -> 5 -> null
-*   L = 2 -> 3 -> null, insert 1, return 1 -> 2 -> 3 -> null
-
-
-
+- L = null, insert 1, return 1 -> null
+- L = 1 -> 3 -> 5 -> null, insert 2, return 1 -> 2 -> 3 -> 5 -> null
+- L = 1 -> 3 -> 5 -> null, insert 3, return 1 -> 3 -> 3 -> 5 -> null
+- L = 2 -> 3 -> null, insert 1, return 1 -> 2 -> 3 -> null
 
 Easy
 
 Linked List
 
-
 ## Assumption
 
 The input node should not be null.
-
 
 ## Algorithm
 
 There are basically two cases:
 
-
-
 1.  insert before the head (head == null || head.value >= value)
 1.  insert into the middle or after the tail: the most important point is to check current.next.val instead of just current.val
 
-
-
-
 ## Solution
 
-
 ### Code
-
 
 ```java
 /**
@@ -80,6 +45,7 @@ There are basically two cases:
  * }
  */
 public class Solution {
+
   public ListNode insert(ListNode head, int value) {
     // Write your solution here
     // Case 1: insert before the head (new node becomes the new head)
@@ -105,13 +71,8 @@ public class Solution {
 }
 ```
 
-
-
 ### Complexity
 
 Time: traversing the entire linked list ⇒ O(n).
 
 Space: constant space ⇒ O(1).
-
-
-<!-- GD2md-html version 1.0β13 -->
